+++ src/Server/Data2.fs
@@ -586,7 +586,7 @@ module Quizzes =
         let TourStatus = "Status"
         let TourQwIdx = "QwIdx"
         let TourQwPartIdx = "QwPartIdx"
-        let TourIsMediaDisplayed = "IsMediaDisplayed"
+        let TourIsQuestionDisplayed = "IsMediaDisplayed"
         let TourStartTime = "StartTime"
         let TourSlip = "Slip"
         let Version = "Version"
@@ -630,9 +630,9 @@ module Quizzes =
 
     let getDescriptor = key >> getItemProjection' dscFields tableName dscReader
 
-    let private tourBuilder name seconds status qwIdx qwPartIdx isMediaDisplayed startTime slip : Domain.QuizTour =
+    let private tourBuilder name seconds status qwIdx qwPartIdx isQuestionDisplayed startTime slip : Domain.QuizTour =
         {Name = name; Seconds = seconds; Status = status |> Option.defaultValue Domain.Announcing;
-            QwIdx = qwIdx; QwPartIdx = qwPartIdx; IsMediaDisplayed = isMediaDisplayed; StartTime = startTime; Slip = slip}
+            QwIdx = qwIdx; QwPartIdx = qwPartIdx; IsQuestionDisplayed = isQuestionDisplayed; StartTime = startTime; Slip = slip}
 
     let private slipChoice =
         function
@@ -646,7 +646,7 @@ module Quizzes =
         <*> (req Fields.TourStatus A.string >- P.enum<Domain.QuizTourStatus>)
         <*> (optDef Fields.TourQwIdx "0" A.number >-> P.int)
         <*> (optDef Fields.TourQwPartIdx "0" A.number >-> P.int)
-        <*> (optDef Fields.TourIsMediaDisplayed false A.bool)
+        <*> (optDef Fields.TourIsQuestionDisplayed false A.bool)
         <*> (opt Fields.TourStartTime (A.nullOr A.string) ??>-> P.dateTime)
         <*> (choice Fields.TourSlip A.docMap slipChoice)
 
@@ -694,8 +694,8 @@ module Quizzes =
                         yield! BuildAttr.optional Fields.TourStartTime ScalarDate tour.StartTime
                         Attr (Fields.TourQwIdx, ScalarInt32 tour.QwIdx)
                         Attr (Fields.TourQwPartIdx, ScalarInt32 tour.QwPartIdx)
-                        if tour.IsMediaDisplayed then
-                            Attr (Fields.TourIsMediaDisplayed, ScalarBool tour.IsMediaDisplayed)
+                        if tour.IsQuestionDisplayed then
+                            Attr (Fields.TourIsQuestionDisplayed, ScalarBool tour.IsQuestionDisplayed)
                     ]
             ])